% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-calculate_credit.R
\name{calculate_credit}
\alias{calculate_credit}
\title{Function to calculate credit dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).}
\usage{
calculate_credit(df)
}
\arguments{
\item{df}{Dataframe containing A-WEAI credit indicator variables.}
}
\value{
A vector recode of WEAI credit dimension of 1 (deficit) or
  0 (non-deficit)
}
\description{
Function to calculate credit dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).
}
\examples{

calculate_credit(df = individual)

}
