% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_wg.R
\name{calculate_domain}
\alias{calculate_domain}
\title{Function to calculate any of the domains of the Washington Group on
Disability indicators.}
\usage{
calculate_domain(df, domain, add = FALSE)
}
\arguments{
\item{df}{Dataframe containing Washington Group on Disability domain-specific
variables.}

\item{domain}{Variable name of domain-specific Washington Group on Disability
indicator.}

\item{add}{Logical. Should the calculated Washington Group on Disability
domain values be added to \code{df}? If TRUE, Washington Group on Disability
results for specific domain will be concatenated to \code{df} using
\code{cbind}. Default is FALSE.}
}
\value{
A dataframe of domain-specific Washington Group on Disability
  indicators.
}
\description{
Function to calculate any of the domains of the Washington Group on
Disability indicators.
}
\examples{

calculate_domain(df = individual, domain = "wg1")

}
