% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-calculate_group.R
\name{calculate_group}
\alias{calculate_group}
\title{Function to calculate group dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).}
\usage{
calculate_group(df)
}
\arguments{
\item{df}{Dataframe containing A-WEAI group indicator variables.}
}
\value{
A vector recode of WEAI group dimension of 1 (deficit) or
  0 (non-deficit)
}
\description{
Function to calculate group dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).
}
\examples{

calculate_group(df = individual)

}
