% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-calculate_mddw.R
\name{calculate_mddw}
\alias{calculate_mddw}
\title{Function to calculate Maternal Dietary Diversity for Women.}
\usage{
calculate_mddw(df, vars = c("mddw1", "mddw2", "mddw3", "mddw4", "mddw5",
  "mddw6"))
}
\arguments{
\item{df}{Dataframe containing Maternal Dietary Diversity for Women
variables.}

\item{vars}{Variables names of MDDW-relevant information.}
}
\value{
A dataframe
}
\description{
Function to calculate Maternal Dietary Diversity for Women.
}
\examples{

calculate_mddw(df = individual)

}
