% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-calculate_production.R
\name{calculate_production}
\alias{calculate_production}
\title{Function to calculate production dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).}
\usage{
calculate_production(df)
}
\arguments{
\item{df}{Dataframe containing A-WEAI production indicator variables.}
}
\value{
A vector recode of WEAI production dimension of 1 (deficit) or
  0 (non-deficit)
}
\description{
Function to calculate production dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).
}
\examples{

calculate_production(df = individual)

}
