% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-calculate_resources.R
\name{calculate_resources}
\alias{calculate_resources}
\title{Function to calculate resources dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).}
\usage{
calculate_resources(df)
}
\arguments{
\item{df}{Dataframe containing A-WEAI resources indicator variables.}
}
\value{
A vector recode of WEAI resources dimension of 1 (deficit) or
  0 (non-deficit)
}
\description{
Function to calculate resources dimension indicators of the Abbreviated
Women's Empowerment in Agriculture Index (A-WEAI).
}
\examples{

calculate_resources(df = individual)

}
