% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate_sers.R
\name{calculate_sers}
\alias{calculate_sers}
\title{Function to calculate Self-evaluated Resiliency Score (SERS)}
\usage{
calculate_sers(df, keep.na = TRUE, response = 3, add = FALSE)
}
\arguments{
\item{df}{Data.frame containing SERS-related variables}

\item{keep.na}{Logical. Apply recoding to NA responses? If TRUE (default),
NA responses will be recoded based on \code{response}. If FALSE, score
will be NA.}

\item{response}{What response to recode NA values. \code{1} for best
possible response; \code{5} for worst possible response; \code{3} for
neutral respnose. Default is \code{3}. Ignored when \code{keep.na} FALSE.}

\item{add}{Logical. Should the calculated SERS values be added to \code{df}?
If TRUE, SERS scores will be concatenated to \code{df} using \code{cbind}.
Default is FALSE.}
}
\value{
If \code{add} is TRUE, a data.frame made out of a column binding of
  \code{df} and the calculated per dimension and overall SERS. If \code{add}
  is FALSE, a SERS data.frame containing per dimension and overall SERS.
}
\description{
Function to calculate Self-evaluated Resiliency Score (SERS)
}
\examples{

calculate_sers(df = individual)

}
