% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_wg.R
\name{calculate_wg}
\alias{calculate_wg}
\title{Function to calculate all domains and overall Washington Group on Disability}
\usage{
calculate_wg(df, add = FALSE)
}
\arguments{
\item{df}{Dataframe containing Washington Group on Disability domain-specific
variables.}

\item{add}{Logical. Should the calculated Washington Group on Disability
values be added to \code{df}? If TRUE, Washington Group on Disability
results will be concatenated to \code{df} using \code{cbind}.
Default is FALSE.}
}
\value{
A dataframe of recoded domain-specific and overall Washington Group
  on Disability values.
}
\description{
Function to calculate all domains and overall Washington Group on Disability
}
\examples{

calculate_wg(df = individual)

}
