% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookCommunity}
\alias{codebookCommunity}
\title{Codebook for raw community dataset from WFP Kenya CSP outcome monitoring
survey}
\format{A data.frame with 13 rows and 3 columns

\describe{
\item{\code{variable}}{Variable name}
\item{\code{question}}{Question/variable description}
\item{\code{choices}}{Choice of responses}
}}
\usage{
codebookCommunity
}
\description{
Codebook for raw community dataset from WFP Kenya CSP outcome monitoring
survey
}
\keyword{datasets}
