% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookHousehold}
\alias{codebookHousehold}
\title{Codebook for raw household dataset from WFP Kenya CSP outcome monitoring
survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 651 rows and 2 columns.}
\usage{
codebookHousehold
}
\description{
A data.frame with 2 columns and 651 rows corresponding to the variables in
the raw \code{household} dataset.
}
\keyword{datasets}
