% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookRoster}
\alias{codebookRoster}
\title{Codebook for raw roster dataset from WFP Kenya CSP outcome monitoring
survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 3 columns.}
\usage{
codebookRoster
}
\description{
A data.frame with 28 rows and 3 columns
}
\keyword{datasets}
