% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community}
\alias{community}
\title{Raw community dataset from WFP Kenya CSP outcome monitoring survey}
\format{A data.frame with 202 rows and 23 columns}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
community
}
\description{
Raw community dataset from WFP Kenya CSP outcome monitoring survey
}
\keyword{datasets}
