% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-wfpKenya_data.R
\name{get_wfp_data}
\alias{get_wfp_data}
\title{Function to pull outcome monitoring data from the ODK server.}
\usage{
get_wfp_data(id, username, password, start = Sys.Date(),
  end = Sys.Date(), filename, remove.prefix = TRUE)
}
\arguments{
\item{id}{Form ID of the different forms designed for the outcome monitoring
study}

\item{username}{ONA server username credentials.}

\item{password}{ONA server password credentials.}

\item{start}{Include data from submission dates after (inclusive) this
start date in export to CSV. Date format <yyyy/MM/dd> and default is
current system date}

\item{end}{Include data from submission dates before (exclusive) this date
in export to CSV. Date format is <yyyy/MM/dd> and default value current
system date}

\item{filename}{Filename to use for data without the CSV file extension.}

\item{remove.prefix}{Logical. If TRUE, remove the ODK group name prefix otherwise
keep. Default is TRUE.}
}
\value{
A data.frame corresponding to dataset corresponding to the form ID
  specified.
}
\description{
Function to pull outcome monitoring data from the ODK server.
}
\examples{
\dontrun{
  ## Get survey data
  get_wfp_data(id = "household",
               username = "validmeasures",
               password = askpass(),
               start = "2019-07-01",
               end = Sys.Date(),
               filename = "household")
}

}
