% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpiYouth}
\alias{gpiYouth}
\title{GPI for youth of household paired dataset}
\format{A data.frame with 818 rows and 8 columns

\describe{
\item{\code{psu}}{Primary sampling unit or village identifier}
\item{\code{hhid}}{Household identifier}
\item{\code{studyGroup}}{Study group; 1 = intervention; 2 = control}
\item{\code{lhz}}{Livelihood zone}
\item{\code{production}}{Production dimension indicator for WEAI}
\item{\code{resources}}{Resources dimension indicator for WEAI}
\item{\code{credit}}{Credit dimension indicator for WEAI}
\item{\code{group}}{Group dimension indicator for WEAI}
}}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
gpiYouth
}
\description{
GPI for youth of household paired dataset
}
\keyword{datasets}
