% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household}
\alias{household}
\title{Raw household dataset from WFP Kenya CSP outcome monitoring survey}
\format{A data.frame with 656 columns and 3057 rows of per household
  observations.}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
household
}
\description{
Raw household dataset from WFP Kenya CSP outcome monitoring survey
}
\keyword{datasets}
