% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual}
\alias{individual}
\title{Raw individual dataset from WFP Kenya CSP outcome monitoring survey}
\format{An object of class \code{data.frame} with 12240 rows and 141 columns.}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
individual
}
\description{
A data.frame with 12240 rows and 141 columns
}
\keyword{datasets}
