% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mddw}
\alias{mddw}
\title{Minimum Dietary Diversity for Women indicators dataset}
\format{A data.frame with 3124 rows and 19 columns

\describe{
\item{\code{psu}}{Primary sampling unit or village identifier}
\item{\code{hhid}}{Household identifier}
\item{\code{sex}}{Sex of respondent; 1 = Male; 2 = Female}
\item{\code{age}}{Age of respondent in years}
\item{\code{respondent}}{Respondent type; 1 = head; 2 = spouse of head;
  3 = male youth household member; 4 = female youth household member}
\item{\code{studyGroup}}{Study group; 1 = intervention; 2 = control}
\item{\code{lhz}}{Livelihood zone}
\item{\code{fg1}}{Food group 1: Grains and tubers}
\item{\code{fg2}}{Food group 2: Pulses}
\item{\code{fg3}}{Food group 3: Nuts and seeds}
\item{\code{fg4}}{Food group 4: Dairy}
\item{\code{fg5}}{Food group 5: Meat, poultry and fish}
\item{\code{fg6}}{Food group 6: Eggs}
\item{\code{fg7}}{Food group 7: Green leafy vegetables}
\item{\code{fg8}}{Food group 8: Other vitamin A rich fruits and vegetables}
\item{\code{fg9}}{Food group 9: Other vegetables}
\item{\code{fg10}}{Food group 10: Other fruits}
\item{\code{fg}}{Total food groups}
\item{\code{mddw}}{Is respondent achieving minimum dietary diversity?
  1 = Yes; 0 = No}
}}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
mddw
}
\description{
Minimum Dietary Diversity for Women indicators dataset
}
\keyword{datasets}
