% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roster}
\alias{roster}
\title{Raw roster dataset from WFP Kenya CSP outcome monitoring survey}
\format{An object of class \code{data.frame} with 16542 rows and 37 columns.}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
roster
}
\description{
A data.frame with 16520 rows and 37 columns
}
\keyword{datasets}
