% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sers}
\alias{sers}
\title{Self-evaluated Resilience Score (SERS) indicator dataset}
\format{A data.frame with 4920 rows and 19 columns

\describe{
\item{\code{psu}}{Primary sampling unit or village identifier}
\item{\code{hhid}}{Household identifier}
\item{\code{sex}}{Sex of respondent; 1 = Male; 2 = Female}
\item{\code{age}}{Age of respondent in years}
\item{\code{respondent}}{Respondent type; 1 = head; 2 = spouse of head;
  3 = male youth household member; 4 = female youth household member}
\item{\code{studyGroup}}{Study group; 1 = intervention; 2 = control}
\item{\code{lhz}}{Livelihood zone}
\item{\code{sers1}}{SERS dimension 1}
\item{\code{sers2}}{SERS dimension 2}
\item{\code{sers3}}{SERS dimension 3}
\item{\code{sers4}}{SERS dimension 4}
\item{\code{sers5}}{SERS dimension 5}
\item{\code{sers6}}{SERS dimension 6}
\item{\code{sers7}}{SERS dimension 7}
\item{\code{sers8}}{SERS dimension 8}
\item{\code{sers9}}{SERS dimension 9}
\item{\code{sers10}}{SERS dimension 10}
\item{\code{score}}{SERS total score}
\item{\code{resilience}}{Is household resilient (score is 1)?
  1 = yes; 2 = no}
}}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
sers
}
\description{
Self-evaluated Resilience Score (SERS) indicator dataset
}
\keyword{datasets}
