% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfpKenya.R
\docType{package}
\name{wfpKenya}
\alias{wfpKenya}
\alias{wfpKenya-package}
\title{The World Food Programme Kenya Country Strategic Plan Outcome Monitoring
aims to assess WFP's progress towards its goals. This package facilitates
the processing, management and analysis of data collected for the outcome
monitoring.}
\description{
The World Food Programme Kenya Country Strategic Plan Outcome Monitoring
aims to assess WFP's progress towards its goals. This package facilitates
the processing, management and analysis of data collected for the outcome
monitoring.
}
