% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wg}
\alias{wg}
\title{Washington Group Disability indicators dataset}
\format{A data.frame with 4920 rows and 36 columns

\describe{
\item{\code{psu}}{Primary sampling unit or village identifier}
\item{\code{hhid}}{Household identifier}
\item{\code{sex}}{Sex of respondent; 1 = Male; 2 = Female}
\item{\code{age}}{Age of respondent in years}
\item{\code{respondent}}{Respondent type; 1 = head; 2 = spouse of head;
  3 = male youth household member; 4 = female youth household member}
\item{\code{studyGroup}}{Study group; 1 = intervention; 2 = control}
\item{\code{lhz}}{Livelihood zone}
\item{\code{wg1.d0}}{No visual disability}
\item{\code{wg1.d1}}{Has any visual disability}
\item{\code{wg1.d2}}{Has moderate or severe visual disability}
\item{\code{wg1.d3}}{Has severe visual disability}
\item{\code{wg2.d0}}{No hearing disability}
\item{\code{wg2.d1}}{Has any hearing disability}
\item{\code{wg2.d2}}{Has moderate or severe hearing disability}
\item{\code{wg2.d3}}{Has severe hearing disability}
\item{\code{wg3.d0}}{Has walking disability}
\item{\code{wg3.d1}}{Has any walking disability}
\item{\code{wg3.d2}}{Has moderate or severe walking disability}
\item{\code{wg3.d3}}{Has severe walking disability}
\item{\code{wg4.d0}}{Has cognitive disability}
\item{\code{wg4.d1}}{Has any cognitive disability}
\item{\code{wg4.d2}}{Has moderate or severe cognitive disability}
\item{\code{wg4.d3}}{Has severe cognitive disability}
\item{\code{wg5.d0}}{Has self-care disability}
\item{\code{wg5.d1}}{Has any self-care disability}
\item{\code{wg5.d2}}{Has moderate or severe self-care disability}
\item{\code{wg5.d3}}{Has severe self-care disability}
\item{\code{wg6.d0}}{Has language disability}
\item{\code{wg6.d1}}{Has any language disability}
\item{\code{wg6.d2}}{Has moderate or severe language disability}
\item{\code{wg6.d3}}{Has severe language disability}
\item{\code{p0}}{No disability in any domain}
\item{\code{p1}}{At least one dimension with disability}
\item{\code{p2}}{At least one dimension with moderate to severe disability}
\item{\code{p3}}{At least one dimension with severe disability}
\item{\code{pm}}{If more than one dimension has disability}
}}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
wg
}
\description{
Washington Group Disability indicators dataset
}
\keyword{datasets}
