% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_wg.R
\name{calculate_domain}
\alias{calculate_domain}
\title{Function to calculate any of the domains of the Washington Group on
Disability indicators.}
\usage{
calculate_domain(df, domain)
}
\arguments{
\item{df}{Dataframe containing Washington Group on Disability domain-specific
variables.}

\item{domain}{Variable name of domain-specific Washington Group on Disability
indicator.}
}
\value{
A dataframe of domain-specific Washington Group on Disability
  indicators.
}
\description{
Function to calculate any of the domains of the Washington Group on
Disability indicators.
}
\examples{

calculate_domain(df = individual, domain = "wg1")

}
