% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-calculate_mddw.R
\name{calculate_mddw}
\alias{calculate_mddw}
\title{Function to recode 18 food items into 10 food groups specific to MDDW.}
\usage{
calculate_mddw(df, vars = c("mddw1", "mddw2", "mddw3", "mddw4", "mddw5",
  "mddw6"))
}
\arguments{
\item{df}{Dataframe containing Maternal Dietary Diversity for Women
variables.}

\item{vars}{Variable names of MDDW-relevant information.}
}
\value{
A dataframe with number of rows equal to number of rows of \code{df}
  and columns for each of the 10 food groups relevent to MDDW recoded into 1
  (consumed) or 0 (not consumed), for total number of food groups consumed,
  and whether a minimum of 5 food groups have been consumed.
}
\description{
Function to recode 18 food items into 10 food groups specific to MDDW.
}
\examples{

calculate_mddw(df = individual)

}
