% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_wg.R
\name{calculate_wg}
\alias{calculate_wg}
\title{Function to calculate all domains and overall Washington Group on Disability}
\usage{
calculate_wg(df)
}
\arguments{
\item{df}{Dataframe containing Washington Group on Disability domain-specific
variables.}
}
\value{
A dataframe of recoded domain-specific and overall Washington Group
  on Disability values.
}
\description{
Function to calculate all domains and overall Washington Group on Disability
}
\examples{

calculate_wg(df = individual)

}
