% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebookRoster}
\alias{codebookRoster}
\title{Codebook for raw roster dataset from WFP Kenya CSP outcome monitoring
survey}
\format{A data.frame with 28 rows and 3 columns
\describe{
\item{\code{variable}}{Variable name}
\item{\code{question}}{Question/variable description}
\item{\code{choices}}{Choice of responses}
}}
\usage{
codebookRoster
}
\description{
Codebook for raw roster dataset from WFP Kenya CSP outcome monitoring
survey
}
\keyword{datasets}
