% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household}
\alias{household}
\title{Raw household dataset from WFP Kenya CSP outcome monitoring survey}
\format{An object of class \code{data.frame} with 3059 rows and 656 columns.}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
household
}
\description{
A data.frame with 656 columns and 3059 rows of per household observations.
}
\keyword{datasets}
