% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual}
\alias{individual}
\title{Raw individual dataset from WFP Kenya CSP outcome monitoring survey}
\format{A data.frame with 4920 rows and 159 columns}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
individual
}
\description{
Raw individual dataset from WFP Kenya CSP outcome monitoring survey
}
\keyword{datasets}
