% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psuData}
\alias{psuData}
\title{Primary sampling unit population dataset}
\format{A data.frame with 202 rows and 2 columns.}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
psuData
}
\description{
Primary sampling unit population dataset
}
\keyword{datasets}
