% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roster}
\alias{roster}
\title{Raw roster dataset from WFP Kenya CSP outcome monitoring survey}
\format{A data.frame with 16518 rows and 39 columns}
\source{
WFP Kenya CSP outcome monitoring survey data collection from
  15-29 August 2019
}
\usage{
roster
}
\description{
Raw roster dataset from WFP Kenya CSP outcome monitoring survey
}
\keyword{datasets}
